#6: Refactor generator code into modules
